refactor(daemon): move three misfiled modules out of agents/
`agents/` should hold agent-domain code. These three were filed there by
accident and belong to other subdomains:

- `config-file-env.ts` -> `shim/` (materializes config-file secrets next to
  the shim's file-sink, which it already imported)
- `commit-message.ts` -> `workspace/` (the workspace git commit-message pass)
- `channel-intro.ts` -> `messages/` (normalized-message intro planning)

Pure `git mv` plus import-path updates across src and test; no symbol renames,
no barrels, no behavior change. Stale path references in nearby comments were
retargeted too.
