+++ packages/daemon/src/session/session-manager.ts
@@ -7,7 +7,7 @@ import { WorkspaceManager } from '../workspace/workspace-manager.js'
 import { initiatorLabel } from '../workspace/session-branch.js'
 import { memoryKindOf, type MemoryProvider, type MemoryScope } from '../agents/memory-provider.js'
 import { agentChildEnv } from '../agents/agent-env.js'
-import { planConfigFiles } from '../agents/config-file-env.js'
+import { planConfigFiles } from '../shim/config-file-env.js'
 import { recalledMemoryBlock, recallQueryFromBlocks, sanitizeRecallRecords } from '../agents/memory-recall.js'
 import type { AcpHost } from '../acp/acp-host.js'
 import type { Agent } from '../agents/agent-schema.js'
