+++ packages/daemon/test/daemon-lifecycle.test.ts
@@ -6,7 +6,7 @@ import { join } from 'node:path'
 import { MAX_TASK_LIST_TASKS } from '@agentconnect.md/protocol'
 import { Daemon } from '../src/daemon.js'
 import { TaskViolationError } from '../src/cp/task-reader.js'
-import { configFilesDir } from '../src/agents/config-file-env.js'
+import { configFilesDir } from '../src/shim/config-file-env.js'
 import { readSkillLedger, skillLedgerLocation } from '../src/skills/skill-install-ledger.js'
 import { sessionKey } from '../src/store/local-store.js'
 import { FakeClock } from './cp/fake-clock.js'
