+++ packages/daemon/test/sandbox.test.ts
@@ -26,7 +26,7 @@ import {
   writeSandboxSettings
 } from '../src/acp/sandbox.js'
 import { claudeInnerSandboxSettings } from '../src/runtime-defs/claude-runtime.js'
-import { clearConfigFiles, configFilesDir, materializeConfigFiles } from '../src/agents/config-file-env.js'
+import { clearConfigFiles, configFilesDir, materializeConfigFiles } from '../src/shim/config-file-env.js'
 
 // Ordinary ACP hosts launch through one SRT provider process with an immutable,
 // daemon-written policy rather than assembling bwrap arguments themselves.
